+++ changelog_entry.yaml
@@ -1,4 +1,5 @@
-- bump: pitch
+- bump: patch
   changes:
     fixed:
     - Colorado House Bill 23-1112 EITC and CTC.
+    - Colorado CTC logic.
