+++ policyengine_us/tests/policy/baseline/gov/states/co/tax/income/credits/ctc/federal_ctc/co_adjusted_federal_ctc.yaml
@@ -1,0 +1,17 @@
+- name: Limiting smaller
+  period: 2022
+  input:
+    state_code: CO
+    co_federal_ctc_maximum: 1_000
+    ctc_limiting_tax_liability: 999
+  output:
+    co_adjusted_federal_ctc: 999
+
+- name: Maximum smaller
+  period: 2022
+  input:
+    state_code: CO
+    co_federal_ctc_maximum: 998
+    ctc_limiting_tax_liability: 1_000
+  output:
+    co_adjusted_federal_ctc: 998
