+++ policyengine_us/tests/policy/baseline/gov/states/co/tax/income/credits/ctc/federal_ctc/co_federal_ctc.yaml
@@ -1,0 +1,199 @@
+- name: No inputs
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 0
+    co_federal_ctc_maximum: 0
+    co_ctc_eligible_children_count: 0
+    tax_unit_earned_income: 0
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 0
+
+- name: No excess
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 1_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 3
+    tax_unit_earned_income: 0
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 0
+
+- name: One child cap at 1,500
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 4_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 1
+    tax_unit_earned_income: 100_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 1_500
+
+- name: Two children cap at 1,500 each
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 6_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 2
+    tax_unit_earned_income: 100_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 3_000
+
+- name: Two children cap max amount
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 4_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 2
+    tax_unit_earned_income: 100_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 2_000
+
+- name: One child with earned income of 10,000 # (10,000 - 2,500) * 0.15
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 4_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 1
+    tax_unit_earned_income: 10_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 0
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 1_125
+
+- name: Three children, child amount over tax
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 10_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 3
+    tax_unit_earned_income: 100_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 1_000
+    employee_medicare_tax: 2_000
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 500
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 4_500
+
+- name: Three children, capped at max amount
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 4_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 3
+    tax_unit_earned_income: 100_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 1_000
+    employee_medicare_tax: 2_000
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 500
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 2_000
+
+- name: Three children, excess child amount under tax
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 10_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 3
+    tax_unit_earned_income: 10_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 0
+    employee_medicare_tax: 0
+    unreported_payroll_tax: 1_000
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 3_000
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 500
+    eitc: 0
+  output:
+    co_additional_federal_ctc: 3_500
+
+- name: Seven children, child amount over tax, reduced by eitc
+  period: 2022
+  input:
+    state_code: CO
+    co_adjusted_federal_ctc: 12_000
+    co_federal_ctc_maximum: 2_000
+    co_ctc_eligible_children_count: 7
+    tax_unit_earned_income: 8_000
+    # SS_ADD_VARIABLES
+    employee_social_security_tax: 1_000
+    employee_medicare_tax: 10_000
+    unreported_payroll_tax: 0
+    self_employment_tax_ald: 0
+    additional_medicare_tax: 0
+    # SS_SUBTRACT_VARIABLES
+    excess_payroll_tax_withheld: 500
+    eitc: 1_000
+  output:
+    co_additional_federal_ctc: 9_500
