+++ policyengine_us/variables/gov/states/co/tax/income/credits/ctc/federal_ctc/co_federal_ctc_child_individual_maximum.py
@@ -16,6 +16,7 @@ class co_federal_ctc_child_individual_maximum(Variable):
         # Colorado Individual Income Tax Filing Guide - Instructions for Select Credits from the DR 0104CR - Line 1 Child Tax Credit
         "https://tax.colorado.gov/sites/tax/files/documents/DR_104_Book_2022.pdf#page=16",
     )
+    defined_for = StateCode.CO
 
     def formula(person, period, parameters):
         age = person("age", period)
