+++ policyengine_us/variables/gov/states/co/tax/income/credits/ctc/federal_ctc/co_federal_ctc_maximum.py
@@ -4,7 +4,7 @@ from policyengine_us.model_api import *
 class co_federal_ctc_maximum(Variable):
     value_type = float
     entity = TaxUnit
-    label = "Maximum CTC replicated to account for the Colorado state CTC child eligibility"
+    label = "Maximum CTC replicated to include the Colorado limitations"
     unit = USD
     documentation = "Maximum value of the Child Tax Credit, before phase-out."
     definition_period = YEAR
@@ -16,5 +16,6 @@ class co_federal_ctc_maximum(Variable):
         # Colorado Individual Income Tax Filing Guide - Instructions for Select Credits from the DR 0104CR - Line 1 Child Tax Credit
         "https://tax.colorado.gov/sites/tax/files/documents/DR_104_Book_2022.pdf#page=16",
     )
+    defined_for = StateCode.CO
 
     adds = ["co_federal_ctc_child_individual_maximum"]
